+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 5.7.8 <small> (2019-09-27)</small>
+
+Update using ```composer update```.
+
+#### Bugfixes
+
+* The correct field name will now be displayed when using wildcard validation rules. 
+
+--------------------------------------------------------
+
 ### 5.7.7 <small> (2019-09-09)</small>
 
 Update using ```composer update```.